+++ src/main/java/com/tech_symfony/movie_booking/api/movie_genre/MovieGenreRepository.java
@@ -1,4 +1,4 @@
-package com.tech_symfony.movie_booking.movie_genre;
+package com.tech_symfony.movie_booking.api.movie_genre;
 
 import com.tech_symfony.movie_booking.model.BaseAuthenticatedRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
