+++ src/main/java/com/tech_symfony/movie_booking/api/movie_status/MovieStatusRepository.java
@@ -1,8 +1,11 @@
-package com.tech_symfony.movie_booking.role;
+package com.tech_symfony.movie_booking.api.movie_status;
 
 import com.tech_symfony.movie_booking.model.BaseAuthenticatedRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
+
 @RepositoryRestResource
-public interface RoleRepository extends BaseAuthenticatedRepository<Role, Integer> {
+public interface MovieStatusRepository extends BaseAuthenticatedRepository<MovieStatus, Integer> {
+
+
 }
