+++ src/main/java/com/tech_symfony/movie_booking/api/order/OrderModelAssembler.java
@@ -1,4 +1,4 @@
-package com.tech_symfony.movie_booking.order;
+package com.tech_symfony.movie_booking.api.order;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 
@@ -15,8 +15,8 @@ class OrderModelAssembler implements RepresentationModelAssembler<Order, EntityM
 		// Unconditional links to single-item resource and aggregate root
 
 		EntityModel<Order> orderModel = EntityModel.of(order,
-				linkTo(methodOn(OrderController.class).one(order.getId())).withSelfRel(),
-				linkTo(methodOn(OrderController.class).all()).withRel("orders"));
+			linkTo(methodOn(OrderController.class).one(order.getId())).withSelfRel(),
+			linkTo(methodOn(OrderController.class).all()).withRel("orders"));
 
 		// Conditional links based on state of the order
