+++ src/main/java/com/tech_symfony/movie_booking/api/order/OrderRepository.java
@@ -1,4 +1,4 @@
-package com.tech_symfony.movie_booking.order;
+package com.tech_symfony.movie_booking.api.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
