+++ src/main/java/com/tech_symfony/movie_booking/api/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.tech_symfony.movie_booking.order;
+package com.tech_symfony.movie_booking.api.order;
 
 enum OrderStatus {
