+++ src/main/java/com/tech_symfony/movie_booking/api/role/RoleRepository.java
@@ -1,11 +1,8 @@
-package com.tech_symfony.movie_booking.movie_status;
+package com.tech_symfony.movie_booking.api.role;
 
 import com.tech_symfony.movie_booking.model.BaseAuthenticatedRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-
 @RepositoryRestResource
-public interface MovieStatusRepository extends BaseAuthenticatedRepository<MovieStatus, Integer> {
-
-
+public interface RoleRepository extends BaseAuthenticatedRepository<Role, Integer> {
 }
