+++ src/main/java/com/tech_symfony/movie_booking/model/BaseAuthenticatedRepository.java
@@ -1,6 +1,5 @@
 package com.tech_symfony.movie_booking.model;
 
-import com.tech_symfony.movie_booking.movie_genre.MovieGenre;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
