+++ src/main/java/com/tech_symfony/movie_booking/model/NamedEntity.java
@@ -3,31 +3,34 @@ package com.tech_symfony.movie_booking.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.MappedSuperclass;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+import lombok.NonNull;
 
 /**
  * Simple JavaBean domain object adds a name property to
  * <code>BaseEntity</code>. Used as
  * a base class for objects needing these properties.
- *
- * 
  */
 @MappedSuperclass
+@Data
 public class NamedEntity extends BaseEntity {
 
-	@Column(name = "name")
+	@Column
+	@NotNull
 	private String name;
 
-	public String getName() {
-		return this.name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	@Override
-	public String toString() {
-		return this.getName();
-	}
+//	public String getName() {
+//		return this.name;
+//	}
+//
+//	public void setName(String name) {
+//		this.name = name;
+//	}
+//
+//	@Override
+//	public String toString() {
+//		return this.getName();
+//	}
 
 }
